cyto/phase5: stabilize ecosystem, migrate orchestration to Spring MVC, and unify observability
- Migrated Expert Review, Notification, Gov Submit, and Crowdfund services to Java 21 / Spring Boot 3.4.1.
- Unified tracing on OTLP and fixed Prometheus/Actuator scrape 404s across the Java stack.
- Implemented Gateway-level Blacklist enforcement via Reactive Redis and fixed 500/401 routing errors.
- Resolved Spring Kafka deserialization crashes by implementing manual Jackson parsing in Java consumers.
- Fixed Python service stability issues (IndentationError, undefined loggers) and updated .gitignore for Java/Node.js/Python artifacts.
- Verified 100% operational status via e2e sanity checks and monitoring target validation.

+++ voting/src/schemas/vote.py
@@ -5,5 +5,5 @@ class VoteRequest(BaseModel):
 
 class VoteResponse(BaseModel):
     target_id: str
-    total_votes: int
+    total_votes: float
     user_vote: int
